+++ internal/docgenerator/docgenerator_test.go
@@ -17,7 +17,7 @@ func TestGenerator_GetNextKey(t *testing.T) {
 		SeedEnd:  1678383842563225000,
 		Template: temp,
 	}
-	for i := 0; i < 10; i++ {
+	for i := int64(0); i < int64(10); i++ {
 		docId, key := g.GetDocIdAndKey(i)
 		log.Println(key, docId)
 		fake := faker.NewWithSeed(rand.NewSource(int64(key)))
